+++ MyShop_Backend/Services/Sizes/ISizeService.cs
@@ -6,7 +6,7 @@ namespace MyShop_Backend.Services.Sizes
 	{
 		Task<IEnumerable<SizeDTO>> GetSizesAsync();
 		Task<SizeDTO> AddSizeAsync(string name);
-		Task<SizeDTO> UpdateSizeAsync(int id, string name);
-		Task DeleteSizeAsync(int id);
+		Task<SizeDTO> UpdateSizeAsync(long id, string name);
+		Task DeleteSizeAsync(long id);
 	}
 }
